Put Export e2e testing into its own helper function
